[DOXIASITETOOLS-295] Rename o.a.m.doxia.siterenderer.Renderer to o.a.m.doxia.siterenderer.SiteRenderer
